+++ .eslintrc
@@ -4,6 +4,9 @@
     "import/extensions": 0,
     "import/no-extraneous-dependencies": 0,
     "import/no-unresolved": [2, { "ignore": ["electron"] }],
-    "linebreak-style": 0
+    "linebreak-style": 0,
+  },
+  "env": {
+    "browser": true
   }
 }
